deep_equation/.ipynb_checkpoints/pacotes-checkpoint.py
import math
import torch
import random
import matplotlib.pyplot as plt
from PIL import Image, ImageOps

from torch import optim
from torch.utils.data import DataLoader
from torch.optim import lr_scheduler
from torch.autograd import Variable
from torch.nn import functional as F
import torch.nn as nn

import torchvision
import torchvision.transforms.functional as F
from torchvision import datasets, models, transforms
from torchvision import datasets
from torchvision.transforms import ToTensor