deep_equation/.ipynb_checkpoints/requirements-checkpoint.txt
tqdm==4.58.0
pytest
Pillow
torchvision